hoverbotpy/pyproject.toml
[tool.poetry]
name = "hoverbotpy"
version = "0.1.0"
description = "Package for hovercraft PIE project."
authors = ["Devlin Ih <dih@olin.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.0.1"
gpiozero = "^1.6.2"
rpi-gpio = "^0.7.1"
lgpio = "^0.0.0.2"
tornado = "^6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hoverbot-discord = "hoverbotpy.hoverbot_discord.bot:main"
